Conflicts:
	bash/colors

+++ deploy.sh
@@ -8,7 +8,7 @@ fi;
 OSX=0
 uname -a | grep Darwin >/dev/null && OSX=1
 
-NOW=$(date +"%m_%d_%Y-%k_%M_%S")
+NOW=$(date +"%m_%d_%Y-%H_%M_%S")
 
 if [[ -f $HOME/.dotfiles/bash/functions ]]; then
 	source "$HOME/.dotfiles/bash/functions"
@@ -24,7 +24,7 @@ mkdir -p $HOME/.old_conf
 
 if [[ -f $HOME/.dotfiles/bash/bashrc ]]; then
 	if [[ -f $HOME/.bashrc ]]; then
-		cp -L $HOME/.bashrc $HOME/.old_conf/bashrc_$NOW && rm $HOME/.bashrc;
+		echo cp -L $HOME/.bashrc $HOME/.old_conf/bashrc_$NOW && rm $HOME/.bashrc;
 	fi;
 	echo "Linking $HOME/.bashrc to $HOME/.dotfiles/bash/bashrc"
 	ln -s $HOME/.dotfiles/bash/bashrc $HOME/.bashrc
@@ -54,7 +54,7 @@ else
 	echo "$HOME/.dotfiles/vim/vimrc: no such file";
 fi
 
-if [[ -f $HOME/.dotfiles/mutt/muttrc ]]; then
+if [[ -f $HOME/.dotfiles/mutt/muttrc &&  OSX ]]; then
 	if [[ -f $HOME/.muttrc ]]; then
 		cp -L $HOME/.muttrc $HOME/.old_conf/muttrc_$NOW && rm $HOME/.muttrc;
 	fi;
@@ -64,7 +64,7 @@ else
 	echo "$HOME/.dotfiles/mutt/muttrc: no such file";
 fi
 
-if [[ -f $HOME/.dotfiles/mutt/mailcap ]]; then
+if [[ -f $HOME/.dotfiles/mutt/mailcap &&  OSX ]]; then
 	if [[ -f $HOME/.mailcap ]]; then
 		cp -L $HOME/.mailcap $HOME/.old_conf/mailcap_$NOW && rm $HOME/.mailcap;
 	fi;
@@ -82,4 +82,54 @@ if [[ -f $HOME/.dotfiles/mutt/urlview &&  OSX ]]; then
 	ln -s $HOME/.dotfiles/mutt/urlview $HOME/.urlview
 else
 	echo "$HOME/.dotfiles/mutt/urlview: no such file";
+fi
+
+if [[ -f $HOME/.dotfiles/afew/config &&  OSX ]]; then
+        if [[ -f $HOME/ ]]; then
+                cp -L $HOME/.config/afew/config $HOME/.old_conf/afew.config_$NOW && rm $HOME/.config/afew/config
+        fi;
+        echo "Linking afew config"
+        ln -s $HOME/.dotfiles/afew/config $HOME/.config/afew/config
+else
+        echo "$HOME/.dotfiles/afew/config: no such file";
+fi
+
+if [[ -f $HOME/.dotfiles/offlineimap/offlineimaprc &&  OSX ]]; then
+        if [[ -f $HOME/ ]]; then
+                cp -L $HOME/.offlineimaprc $HOME/.old_conf/afew.config_$NOW && rm $HOME/.offlineimaprc
+        fi;
+        echo "Linking offlineimap config"
+        ln -s $HOME/.dotfiles/offlineimap/offlineimaprc $HOME/.offlineimaprc
+else
+        echo "$HOME/.dotfiles/offlineimap/offlineimaprc: no such file";
+fi
+
+if [[ -f $HOME/.dotfiles/notmuch/notmuch-config &&  OSX ]]; then
+        if [[ -f $HOME/ ]]; then
+                cp -L $HOME/.notmuch-config $HOME/.old_conf/notmuch_$NOW && rm $HOME/.notmuch-config
+        fi;
+        echo "Linking notmuch config"
+        ln -s $HOME/.dotfiles/notmuch/notmuch-config $HOME/.notmuch-config
+else
+        echo "$HOME/.dotfiles/notmuch/notmuch-config: no such file";
+fi
+
+if [[ -f $HOME/.dotfiles/notmuch/post-new &&  OSX ]]; then
+        if [[ -f $HOME/ ]]; then
+                cp -L $HOME/.mail/.notmuch/hooks/post-new $HOME/.old_conf/notmuch_post-new && rm $HOME/.mail/.notmuch/hooks/post-new
+        fi;
+        echo "Linking notmuch post-new config"
+        ln -s $HOME/.dotfiles/notmuch/post-new $HOME/.mail/.notmuch/hooks/post-new
+else
+        echo "$HOME/.dotfiles/notmuch/post-new: no such file";
+fi
+
+if [[ -f $HOME/.dotfiles/notmuch/pre-new &&  OSX ]]; then
+        if [[ -f $HOME/ ]]; then
+                cp -L $HOME/.mail/.notmuch/hooks/pre-new $HOME/.old_conf/notmuch_pre-new && rm $HOME/.mail/.notmuch/hooks/pre-new
+        fi;
+        echo "Linking notmuch pre-new config"
+        ln -s $HOME/.dotfiles/notmuch/pre-new $HOME/.mail/.notmuch/hooks/pre-new
+else
+        echo "$HOME/.dotfiles/notmuch/pre-new: no such file";
 fi
